feat: se refactorizó el componente habilitar-reporte
