+++ src/app/@core/models/interfaces/DataRequest.interface.ts
@@ -1,0 +1,6 @@
+export interface DataRequest {
+  Data:    any[] | any;
+  Message: string;
+  Status:  string;
+  Success: boolean;
+}
