+++ src/app/pages/plan/habilitar-reporte/tabla-unidades/tabla-unidades.component.ts
@@ -10,22 +10,7 @@ import { RequestManager } from '../../../services/requestManager';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
 import Swal from 'sweetalert2';
-
-export interface Unidades {
-  CorreoElectronico: string;
-  DependenciaTipoDependencia: string;
-  Id: number;
-  Nombre: string;
-  TelefonoDependencia: string;
-  TipoDependencia: any;
-  iconSelected: string;
-  posicion: string;
-}
-
-export interface Unidad {
-  Id: string;
-  Nombre: string;
-}
+import { Unidades } from '../utils/models-habilitar-reporte';
 
 @Component({
   selector: 'app-tabla-unidades',
