+++ src/app/pages/plan/habilitar-reporte/utils/constantes.ts
@@ -1,0 +1,4 @@
+export const PROCESO_FUNCIONAMIENTO_FORMULACION = 'formulacion';
+export const PROCESO_FUNCIONAMIENTO_SEGUIMIENTO = 'seguimiento';
+export const PROCESO_INVERSION_FORMULACION = 'formulaciones';
+export const PROCESO_INVERSION_SEGUIMIENTO = 'seguimientos';
